requirements.txt
presidio-analyzer>=2.2.355,<3
presidio-anonymizer>=2.2.355,<3
spacy>=3.8,<4
PyPDF2>=3.0.1
python-docx>=1.1.2
pyinstaller>=6.11.1
pyinstaller-hooks-contrib>=2025.8
altgraph>=0.17.4
macholib>=1.16.3; platform_system == "Darwin"
